remove tree app
