+++ backend/sync-helper/config/settings.py
@@ -40,7 +40,6 @@ INSTALLED_APPS = [
     
     'rest_framework',
     'chromium',
-    'tree',
 ]
 
 MIDDLEWARE = [
