--- backend/sync-helper/tree/models.py
@@ -1,2 +1,0 @@
-from django.db import models
-
